Take advantage of the .state file

+++ lambda/utilities/delete-hello-world.sh
@@ -21,6 +21,12 @@ if ! . $UTILIRY_DIRECTORY/check-variables.sh ; then
   exit 1
 fi
 
+if ! [ -f $UTILIRY_DIRECTORY/.hello-world ] ; then 
+  echo 'State file is not present. Lambda not deleted.'
+  echo 'If you know the Lambda is created, touch utilities/.hello-world and re-run this script.' 
+  exit 0
+fi 
+
 echo 'Attempting to destroy AWS Lambda with name "'$AWS_LAMBDA_NAME'" and AWS Role "'$AWS_LAMBDA_ROLE_WORKSHOP'".' 
 
 if aws lambda delete-function \
